Issue #1045 - Update projection
- MOve TileProjection and MercatorProjection to their own headers
- Move Projections to projection package/directory
- Adapt includes
- Fixed builds
