+++ Tests/src/MercatorProjection.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
 
-#include <osmscout/util/Projection.h>
+#include "osmscout/projection/MercatorProjection.h"
 
 #include <TestMain.h>
