+++ libosmscout-map-gdi/src/osmscoutmapgdi/MapPainterGDI.cpp
@@ -349,8 +349,7 @@ namespace osmscout {
 
       result.emplace_back();
       result.back().glyph.character = WStringToUTF8String(buffer.data());
-      result.back().position.SetX(horizontalOffset);
-      result.back().position.SetY(0);
+      result.back().position=Vertex2D(horizontalOffset,0);
 
       Gdiplus::RectF bb;
