+++ libosmscout-map/include/osmscoutmap/MapPainter.h
@@ -40,9 +40,10 @@
 
 #include <osmscoutmap/MapData.h>
 
+#include <osmscout/projection/Projection.h>
+
 #include <osmscout/util/Breaker.h>
 #include <osmscout/util/Geometry.h>
-#include <osmscout/util/Projection.h>
 #include <osmscout/util/Transformation.h>
 
 #include <osmscout/system/Compiler.h>
