+++ libosmscout/include/meson.build
@@ -4,6 +4,9 @@ osmscoutHeader = [
             'osmscout/CoreImportExport.h',
             'osmscout/ost/Parser.h',
             'osmscout/ost/Scanner.h',
+            'osmscout/projection/MercatorProjection.h',
+            'osmscout/projection/Projection.h',
+            'osmscout/projection/TileProjection.h',
             'osmscout/system/Assert.h',
             'osmscout/system/Compiler.h',
             'osmscout/system/Math.h',
@@ -37,7 +40,6 @@ osmscoutHeader = [
             'osmscout/util/PolygonCenter.h',
             'osmscout/util/ProcessingQueue.h',
             'osmscout/util/Progress.h',
-            'osmscout/util/Projection.h',
             'osmscout/util/ScopeGuard.h',
             'osmscout/util/ScreenBox.h',
             'osmscout/util/StopClock.h',
