+++ libosmscout/include/osmscout/projection/Projection.h
@@ -29,7 +29,6 @@
 #include <osmscout/util/GeoBox.h>
 #include <osmscout/util/Magnification.h>
 #include <osmscout/util/ScreenBox.h>
-#include <osmscout/util/Tiling.h>
 
 #include <osmscout/system/SSEMathPublic.h>
 
@@ -78,8 +77,8 @@ namespace osmscout {
       int                count=0;
       ALIGN16_BEG double lon[2] ALIGN16_END;
       ALIGN16_BEG double lat[2] ALIGN16_END;
-      double*            xPointer[2]=NULL;
-      double*            yPointer[2]=null;
+      double*            xPointer[2]=nullptr;
+      double*            yPointer[2]=nullptr;
 #endif
 
     public:
@@ -342,269 +341,6 @@ namespace osmscout {
 
     friend class BatchTransformer;
   };
-
-  /**
-   * Mercator projection that tries to render the resulting map in the same
-   * physical size on all devices. If the physical DPI of the device is
-   * correctly given, objects on any device has the same size. Bigger devices
-   * will show "more" map thus.
-   *
-   * Scale is calculated based on the assumption that the original OpenStreetMap
-   * tiles were designed for 96 DPI displays.
-   *
-   */
-  class OSMSCOUT_API MercatorProjection : public Projection
-  {
-  protected: // Because the OpenGLProjection inherits from this projection
-    bool   valid=false;    //!< projection is valid
-
-    double latOffset=0.0;  //!< Absolute and untransformed screen position of lat coordinate
-    double angleSin=0.0;
-    double angleCos=0.0;
-    double angleNegSin=0.0;
-    double angleNegCos=0.0;
-
-    double scale=1.0;
-    double scaleGradtorad=0.0 ; //!< Precalculated scale*Gradtorad
-
-    double scaledLatDeriv=0.0; //!< precalculated derivation of "latToYPixel" function in projection
-                           //!< center scaled by gradtorad * scale
-    bool   useLinearInterpolation=false; //!< switch to enable linear interpolation of latitude to pixel computation
-
-  public:
-    static const double MaxLat;
-    static const double MinLat;
-    static const double MaxLon;
-    static const double MinLon;
-
-    MercatorProjection() = default;
-    MercatorProjection(const MercatorProjection&) = default;
-    MercatorProjection(MercatorProjection&&) = default;
-    MercatorProjection& operator=(const MercatorProjection&) = default;
-    MercatorProjection& operator=(MercatorProjection&&) = default;
-    ~MercatorProjection() override = default;
-
-    bool CanBatch() const override
-    {
-      return false;
-    }
-
-    bool IsValid() const override
-    {
-      return valid;
-    }
-
-    bool IsValidFor(const GeoCoord& coord) const override
-    {
-      return coord.GetLat() >= MinLat && coord.GetLat() <= MaxLat &&
-             coord.GetLon() >= MinLon && coord.GetLon() <= MaxLon;
-    }
-
-    bool Set(const GeoCoord& coord,
-                    const Magnification& magnification,
-                    size_t width,size_t height)
-    {
-      return Set(coord,0.0,magnification,GetDPI(),width,height);
-    }
-
-    bool Set(const GeoCoord& coord,
-                    double angle,
-                    const Magnification& magnification,
-                    size_t width, size_t height)
-    {
-      return Set(coord,angle,magnification,GetDPI(),width,height);
-    }
-
-    bool Set(const GeoCoord& coord,
-                    const Magnification& magnification,
-                    double dpi,
-                    size_t width, size_t height)
-    {
-      return Set(coord,0.0,magnification,dpi,width,height);
-    }
-
-    /**
-     * Setup projection parameters.
-     *
-     * Return true on success,
-     * false if arguments are not valid for Mercator projection,
-     * projection parameters are unchanged in such case.
-     *
-     * Angle is in radians ([0..2*PI[)
-     *
-     * Note that coord (center) have to be valid coordinate
-     * in Mercator projection. But it is possible setup dimensions
-     * (width and height) that projection will cover area bigger
-     * than the one valid for Mercator projection. Bounding box
-     * is adjusted then to be valid for projection.
-     *
-     * In code:
-     *
-     *   projection.GetDimensions(bbox);
-     *   projection.GeoToPixel(bbox.GetMinCoord(),x,y)
-     *
-     * may be x >= 0
-     */
-    bool Set(const GeoCoord& coord,
-             double angle,
-             const Magnification& magnification,
-             double dpi,
-             size_t width, size_t height);
-
-    bool PixelToGeo(double x, double y,
-                    GeoCoord& coord) const override;
-
-    bool GeoToPixel(const GeoCoord& coord,
-                    Vertex2D& pixel) const override;
-
-    bool Move(double horizPixel,
-              double vertPixel);
-
-    bool MoveUp(double pixel)
-    {
-      return Move(0,pixel);
-    }
-
-    bool MoveDown(double pixel)
-    {
-      return Move(0,-pixel);
-    }
-
-    bool MoveLeft(double pixel)
-    {
-      return Move(-pixel,0);
-    }
-
-    bool MoveRight(double pixel)
-    {
-      return Move(pixel,0);
-    }
-
-    [[nodiscard]] bool IsLinearInterpolationEnabled() const
-    {
-      return useLinearInterpolation;
-    }
-
-    /**
-     * Switch to enable/disable linear interpolation of latitude to pixel computation.
-     * It speedup GeoToPixel calculation with fractional error on small render area.
-     */
-    void SetLinearInterpolationUsage(bool useLinearInterpolation)
-    {
-      this->useLinearInterpolation=useLinearInterpolation;
-    }
-
-  protected:
-    void GeoToPixel(const BatchTransformer& transformData) const override;
-  };
-
-
-  /**
-   * Mercator projection as used by the OpenStreetMap tile rendering code.
-   *
-   * The TileProjection simplifies the general Mercator projection code to
-   * make sure that there are no effects based on rounding errors or similar.
-   */
-  class OSMSCOUT_API TileProjection : public Projection
-  {
-  private:
-    bool   valid=false;          //!< projection is valid
-
-    double lonOffset=0.0;
-    double latOffset=0.0;
-    double scale=1.0;
-    double scaleGradtorad=0.0; //!< Precalculated scale*Gradtorad
-
-    double scaledLatDeriv=0.0; //!< precalculated derivation of "latToYPixel" function in projection
-                           //!< center scaled by gradtorad * scale
-    bool   useLinearInterpolation=false; //!< switch to enable linear interpolation of latitude to pixel computation
-
-#ifdef OSMSCOUT_HAVE_SSE2
-    //some extra vars for special sse needs
-      v2df              sse2LonOffset;
-      v2df              sse2LatOffset;
-      v2df              sse2Scale;
-      v2df              sse2ScaleGradtorad;
-      v2df              sse2Height;
-#endif
-
-  protected:
-    virtual bool SetInternal(double lonMin,double latMin,
-                             double lonMax,double latMax,
-                             const Magnification& magnification,
-                             double dpi,
-                             size_t width,size_t height);
-
-  public:
-    TileProjection() = default;
-    TileProjection(const TileProjection&) = default;
-    TileProjection(TileProjection&&) = default;
-    TileProjection& operator=(const TileProjection&) = default;
-    TileProjection& operator=(TileProjection&&) = default;
-    ~TileProjection() override = default;
-
-    bool CanBatch() const override
-    {
-      return true;
-    }
-
-    bool IsValid() const override
-    {
-      return valid;
-    }
-
-    bool IsValidFor(const GeoCoord& coord) const override
-    {
-      return coord.GetLat() >= -85.0511 && coord.GetLat() <= +85.0511 &&
-             coord.GetLon() >= -180.0   && coord.GetLon() <= +180.0;
-    }
-
-    bool Set(const OSMTileId& tile,
-                    const Magnification& magnification,
-                    size_t width, size_t height)
-    {
-      return Set(tile,
-                 magnification,
-                 GetDPI(),
-                 width,
-                 height);
-    }
-
-    bool Set(const OSMTileId& tile,
-             const Magnification& magnification,
-             double dpi,
-             size_t width, size_t height);
-
-    bool Set(const OSMTileIdBox& tileBox,
-             const Magnification& magnification,
-             double dpi,
-             size_t width, size_t height);
-
-    bool PixelToGeo(double x, double y,
-                    GeoCoord& coord) const override;
-
-    bool GeoToPixel(const GeoCoord& coord,
-                    Vertex2D& pixel) const override;
-
-    [[nodiscard]] bool IsLinearInterpolationEnabled() const
-    {
-      return useLinearInterpolation;
-    }
-
-    /**
-     * Switch to enable/disable linear interpolation of latitude to pixel computation.
-     * It speedup GeoToPixel calculation with fractional error on small render area.
-     */
-    void SetLinearInterpolationUsage(bool b)
-    {
-      useLinearInterpolation = b;
-    }
-
-  protected:
-
-    void GeoToPixel(const BatchTransformer& transformData) const override;
-  };
-
 }
 
 #endif
