+++ libosmscout/include/osmscout/util/Transformation.h
@@ -24,9 +24,10 @@
 
 #include <osmscout/CoreImportExport.h>
 
+#include <osmscout/projection/Projection.h>
+
 #include <osmscout/util/Geometry.h>
 #include <osmscout/util/Logger.h>
-#include <osmscout/util/Projection.h>
 
 #include <osmscout/system/Assert.h>
 #include <osmscout/system/Math.h>
